pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-llm-cache"
version = "0.1.0"
description = "LLM caching helpers and pytest plugin for Gemini/OpenAI workflows"
readme = "README.md"
authors = [{name = "jjeaby@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "google-generativeai",
    "openai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = "src"

[project.entry-points."pytest11"]
llm_cache = "pytest_llm_cache.pytest_llm_cache"
